Return YAS orientation to 5
The YAS mag sensor is an interesting case. With the position set to 0, the compass is kind-of "on" at bootup, but after calibration, is way off. With it set to 5 like factory, it is way off at bootup, but reasonably accurate after calibration. Ideally, we need to implement some mechanism to store and load calibrations.